WIP: move code out of kubelet.go
